+++ CHANGELOG.md
@@ -6,6 +6,11 @@ These are the changes that have not been released yet. The release notes have no
 
 **FEATURE ENCHANCEMENTS:**
 
+- Re-add configuration file `podmod.conf`.
+    - Specify kernel module parameters in configuration file.
+    - Specify container image build parameters in configuration file.
+    - Tag images with module version from configuration file.
+
 - Add `--no-prune` option to `build`.
 
 - Add `--idempotent` option to `build`, `load`, and `unload`.